# feed-reader/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Feed Reader App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /login:
    # binds a127 app logic to a route
    x-swagger-router-controller: login
    get:
      description: Returns '200' to the caller if params hase "username" of type String
      # used as the method name of the controller
      operationId: login
      parameters:
        - name: username
          in: query
          description: allows a simulated login, but no real authentication
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LoginResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /registerFeed:
      x-swagger-router-controller: feed.methods
      post:
        description: Add new feed
        operationId: registerFeed
        parameters:
          - name: feed
            in: body
            description: object to create newfeed
            schema:
              $ref: "#/definitions/Feed"
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/RegisterFeedResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /feeds/fetch/{userName}:
    x-swagger-router-controller: feed.methods
    get:
      description: Fetch feed definitions per user from the database.
      operationId: fetchFeeds
      parameters:
        - in: path
          name: userName
          required: true
          type: string
          description: The user name.
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /feeds/{feedId}:
    x-swagger-router-controller: feed.methods
    delete:
      description: Delete feeds by "_id" from the database.
      operationId: deleteFeed
      parameters:
        - in: path
          name: feedId
          required: true
          type: string
          description: Feed _id.
      responses:
        200:
          description: Success
          schema:
            type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Feed:           # <----------
    type: object
    required:
      - userName
      - title
    properties:
      userName:
        type: string
      title:
        type: string
      URL:
        type: string
  LoginResponse:
    description: User was logged in.
  RegisterFeedResponse:
    description: Feed was created.
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string